Wired KBase cells so that deleting them (with the delete buttons) should also delete any associated job.

+++ src/notebook/ipython_profiles/profile_narrative/kbase_templates/static/kbase/js/widgets/narrative_core/kbaseNarrativeWorkspace.js
@@ -110,16 +110,6 @@
                 this)
             );
 
-            // unused for now.
-            // maybe update to modify vis of KBase widgets.
-            // but nuking for the moment.
-            // $(document).on('servicesUpdated.Narrative',
-            //     $.proxy(function(event, serviceSet) {
-            //         console.log("listing services!");
-            //     },
-            //     this)
-            // );
-
             $(document).on('narrativeDataQuery.Narrative', 
                 $.proxy(function(e, callback) {
                     var objList = this.getNarrativeDependencies();
@@ -1256,7 +1246,71 @@
          * @private
          */
         deleteCell: function(index) {
-            IPython.notebook.delete_cell(index);
+            if (index !== undefined && index !== null) {
+                var cell = IPython.notebook.get_cell(index);
+                if (cell) {
+                    var okayToDelete = false;
+                    // if it's a kbase method or app cell, trigger a popup
+                    if (cell.metadata[this.KB_CELL]) {
+                        widget = null; // default is app cell
+                        var state = 'input'; // default is input... also doubles as a proxy for output cells
+                        if (this.isFunctionCell(cell))
+                            widget = 'kbaseNarrativeMethodCell';
+                        else if (this.isAppCell(cell))
+                            widget = 'kbaseNarrativeAppCell';
+                        if (widget)
+                            state = $(cell.element).find('div[id^=kb-cell-]')[widget]('getRunningState');
+
+                        if (state === 'input')
+                            okayToDelete = true;
+                        else {
+                            // if it's running, say so, and say it'll stop and delete the job
+                            // if it's done, say it'll clear the associated job, but won't delete data
+                            // if it's error, say it'll delete the assoc'd job
+
+                            var stateWarning = 'Deleting this cell will also delete any associated job. ' +
+                                               'Any generated data will be retained. Continue?';
+
+                            console.log(stateWarning);
+                            okayToDelete = true;
+                            // switch(state) {
+                            //     case 'running':
+                            //         // set some text
+                            //         stateWarning = 'This cell appears to have a running job associated with it. ' +
+                            //                        'Deleting this cell will also stop and delete the running job. ' +
+                            //                        'Any generated data will not be deleted. Continue?';
+                            //         break;
+                            //     case 'error':
+                            //         // set some text
+                            //         stateWarning = 'This cell appears to have produced an error while running. ' +
+                            //                        'Deleting this cell will also stop and delete the associated job. ' +
+                            //                        'Any generated data will be maintained. Continue?';
+                            //         break;
+                            //     case 'done':
+                            //         // set some text
+                            //         stateWarning = 'This cell has finished running but may have a job still associated with it. ' +
+                            //                        'Deleting this cell will also delete that job, but not any generated data. ' +
+                            //                        'Continue?';
+                            //         break;
+                            //     default:
+                            //         // set no text
+                            //         stateWarning = 'Deleting this cell will also delete any associated job. Any generated data will be retained. Continue?';
+                            //         break;
+                            // }
+                        }
+                        if (okayToDelete) {
+                            // try to kill the job.
+                            var cellId = $(cell.element).find('div[id^=kb-cell-]').attr('id');
+                            this.trigger('cancelJobCell.Narrative', cellId, false);
+                        }
+                    }
+                    else {
+                        okayToDelete = true;
+                    }
+                    if (okayToDelete)
+                        IPython.notebook.delete_cell(index);
+                }
+            }
         },
 
         /**
